+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "auditore-forum",
-  "version": "1.1.0",
+  "version": "1.2",
   "private": true,
   "scripts": {
     "dev": "next dev",
